requirements.txt
fastapi
uvicorn
langgraph
langchain
langchain-openai
sentence-transformers
qdrant-client
tavily-python
dspy-ai
python-dotenv
datasets
scikit-learn
beautifulsoup4
requests
fastmcp
mcp
langchain_groq
openai
groq
sympy
streamlit